# docs/swagger.yaml
basePath: /gsafety
definitions:
...
  request.FacilityAdd:
    properties:
      code:
        description: 新增参数：设备唯一标识码
        type: string
      name:
        description: 新增参数：设备名
        type: string
      status:
        description: 新增参数：设备状态 0,1,2 <==> 在线 离线 告警
        type: string
      type:
        description: 新增参数：设备类型 0,1,2,3,4,5 <==> 电子围栏、火灾报警器、喇叭、CCTV、警报器、门禁
        type: string
    type: object
  request.FacilityUpdate:
    properties:
      alarmTime:
        description: 报警时间
        type: string
      code:
        description: 更新参数：设备唯一标识码
        type: string
      id:
        description: 更新参数：主键ID
        example: "0"
        type: string
      name:
        description: 更新参数：设备名
        type: string
      status:
        description: 更新参数：设备状态 0,1,2 <==> 在线 离线 告警
        type: string
      type:
        description: 更新参数：设备类型 0,1,2,3,4,5 <==> 电子围栏、火灾报警器、喇叭、CCTV、警报器、门禁
        type: string
    type: object
...
  response.Email:
    properties:
      body:
        description: 邮件内容
        type: string
      subject:
        description: 邮件标题
        type: string
      to:
        description: 邮件发送给谁
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 智慧园区demo后端服务api
  title: Swagger Gsafety demo API
  version: 0.0.1
paths:
...
  /email/emailTest:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"发送成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 发送测试邮件
      tags:
      - System
  /email/sendEmail:
    post:
      parameters:
      - description: 发送邮件必须的参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/response.Email'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"发送成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 发送邮件
      tags:
      - System
  /facility/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 添加设备
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.FacilityAdd'
      produces:
      - application/json
      responses:
        "200":
          description: okay
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 添加设备
      tags:
      - FacilityApi
  /facility/get:
    get:
      parameters:
      - description: 通过设备id获取设备详情
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: okay
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 获取设备详情
      tags:
      - FacilityApi
  /facility/getList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询参数：设备唯一标识码
        in: query
        name: code
        type: string
      - description: 查询参数：报警结束
        in: query
        name: endTime
        type: string
      - description: 关键字 不做特定约束 根据实际接口设计选用 可用来做查询过滤、排序等，做排序时需满足param1-desc,param2-asc,param3-asc这种格式
        in: query
        name: keyword
        type: string
      - description: 查询参数：设备名
        in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 查询参数：报警起始时间
        in: query
        name: startTime
        type: string
      - description: 查询参数：设备状态 用,隔开 0,1,2 <==> 在线 离线 告警
        in: query
        name: status
        type: string
      - description: 查询参数：设备类型 用,隔开 0,1,2,3,4,5 <==> 电子围栏、火灾报警器、喇叭、CCTV、警报器、门禁
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取设备列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      summary: 分页获取设备列表
      tags:
      - FacilityApi
  /facility/remove:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 通过设备id删除设备
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: okay
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 删除设备
      tags:
      - FacilityApi
  /facility/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新设备
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.FacilityUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: okay
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 更新设备
      tags:
      - FacilityApi
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"